# docs/CHATGPT_ACTIONS_OPENAPI.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DarkLock On-The-Go Status API",
    "version": "1.0.0",
    "description": "Token-protected status endpoints for ChatGPT Actions."
  },
  "servers": [
    {
      "url": "https://darklock.net/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/chatgpt/health": {
      "get": {
        "operationId": "getDarklockHealth",
        "summary": "Get platform health snapshot",
        "responses": {
          "200": {
            "description": "Health snapshot"
          },
          "401": {
            "description": "Unauthorized"
          },
          "503": {
            "description": "Unavailable"
          }
        }
      }
    },
    "/chatgpt/services": {
      "get": {
        "operationId": "getDarklockServices",
        "summary": "Get key service reachability",
        "responses": {
          "200": {
            "description": "Service reachability status"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/chatgpt/status": {
      "get": {
        "operationId": "getDarklockStatusSummary",
        "summary": "Get consolidated status summary with runtime bot signal",
        "responses": {
          "200": {
            "description": "Status summary payload"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/chatgpt/info": {
      "get": {
        "operationId": "getDarklockInfo",
        "summary": "Get quick links and environment info",
        "responses": {
          "200": {
            "description": "Quick reference info"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    }
  }
}